+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.1.0
 
  - Add `not-contains` assertion on `stdout` and `stderr`
+ - Add validation for invalid data types in `stdout` and `stderr` assertions
+ - More logging for `--verbose` option on the `test` command
 
 # v1.0.1